+++ bsps/arm/raspberrypi/console/fbcons.c
@@ -100,7 +100,7 @@ static void fbcons_write_support_polled(
   size_t      len
 )
 {
-  int nwrite = 0;
+  size_t nwrite = 0;
 
   /*
    * poll each byte in the string out of the port.
